+++ src/main/java/fr/dedoublonneur/workflow/CleanupNotAllowedException.java
@@ -1,0 +1,12 @@
+package fr.dedoublonneur.workflow;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.CONFLICT)
+public class CleanupNotAllowedException extends RuntimeException {
+
+    public CleanupNotAllowedException(String message) {
+        super(message);
+    }
+}